requirements.txt
Django==5.2.11
djangorestframework==3.16.1
psycopg2-binary==2.9.10
gunicorn==22.0.0
django-cors-headers==4.4.0
